requirements.txt
numpy
google-generativeai
rich